Add util fn to make titles consistent
